LayoutTests/fast/shapes/parsing/parsing-shape-outside-expected.txt

Testing parsing of the -webkit-shape-outside property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getCSSText("-webkit-shape-outside", "none") is "none"
PASS getComputedStyleValue("-webkit-shape-outside", "none") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px)") is "inset(10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px)") is "inset(10px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px 9px)") is "inset(10px 9px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 9px)") is "inset(10px 9px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px 9px 8px)") is "inset(10px 9px 8px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 9px 8px)") is "inset(10px 9px 8px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px 20px 30px 40px)") is "inset(10px 20px 30px 40px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 20px 30px 40px)") is "inset(10px 20px 30px 40px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px 20px 30px 20px)") is "inset(10px 20px 30px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 20px 30px 20px)") is "inset(10px 20px 30px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px 20px 10px 20px)") is "inset(10px 20px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 20px 10px 20px)") is "inset(10px 20px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px)") is "inset(10px round 9px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px)") is "inset(10px round 9px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 0px / 0px 10px)") is "inset(10px round 0px / 0px 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 0px / 0px 10px)") is "inset(10px round 0px / 0px 10px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px / 9px)") is "inset(10px round 9px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px / 9px)") is "inset(10px round 9px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px / 8px)") is "inset(10px round 9px / 8px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px / 8px)") is "inset(10px round 9px / 8px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px / 8px 7px)") is "inset(10px round 9px / 8px 7px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px / 8px 7px)") is "inset(10px round 9px / 8px 7px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px / 8px 7px 6px)") is "inset(10px round 9px / 8px 7px 6px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px / 8px 7px 6px)") is "inset(10px round 9px / 8px 7px 6px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px / 8px 7px 6px 5px)") is "inset(10px round 9px / 8px 7px 6px 5px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px / 8px 7px 6px 5px)") is "inset(10px round 9px / 8px 7px 6px 5px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px / 7px)") is "inset(10px round 9px 8px / 7px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px / 7px)") is "inset(10px round 9px 8px / 7px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px / 7px 6px)") is "inset(10px round 9px 8px / 7px 6px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px / 7px 6px)") is "inset(10px round 9px 8px / 7px 6px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px / 7px 6px 5px)") is "inset(10px round 9px 8px / 7px 6px 5px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px / 7px 6px 5px)") is "inset(10px round 9px 8px / 7px 6px 5px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px / 7px 6px 5px 4px)") is "inset(10px round 9px 8px / 7px 6px 5px 4px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px / 7px 6px 5px 4px)") is "inset(10px round 9px 8px / 7px 6px 5px 4px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px 7px / 6px)") is "inset(10px round 9px 8px 7px / 6px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px 7px / 6px)") is "inset(10px round 9px 8px 7px / 6px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px 7px / 6px 5px)") is "inset(10px round 9px 8px 7px / 6px 5px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px 7px / 6px 5px)") is "inset(10px round 9px 8px 7px / 6px 5px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px 7px / 6px 5px 4px)") is "inset(10px round 9px 8px 7px / 6px 5px 4px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px 7px / 6px 5px 4px)") is "inset(10px round 9px 8px 7px / 6px 5px 4px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px 7px / 6px 5px 4px 3px)") is "inset(10px round 9px 8px 7px / 6px 5px 4px 3px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px 7px / 6px 5px 4px 3px)") is "inset(10px round 9px 8px 7px / 6px 5px 4px 3px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px 7px 6px / 5px)") is "inset(10px round 9px 8px 7px 6px / 5px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px 7px 6px / 5px)") is "inset(10px round 9px 8px 7px 6px / 5px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px 7px 6px / 5px 4px)") is "inset(10px round 9px 8px 7px 6px / 5px 4px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px 7px 6px / 5px 4px)") is "inset(10px round 9px 8px 7px 6px / 5px 4px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px 7px 6px / 5px 4px 3px)") is "inset(10px round 9px 8px 7px 6px / 5px 4px 3px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px 7px 6px / 5px 4px 3px)") is "inset(10px round 9px 8px 7px 6px / 5px 4px 3px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 9px 8px 7px 6px / 5px 4px 3px 2px)") is "inset(10px round 9px 8px 7px 6px / 5px 4px 3px 2px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 9px 8px 7px 6px / 5px 4px 3px 2px)") is "inset(10px round 9px 8px 7px 6px / 5px 4px 3px 2px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 10px 20px 30px 20px)") is "inset(10px round 10px 20px 30px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 10px 20px 30px 20px)") is "inset(10px round 10px 20px 30px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 10px 20px 10px 20px)") is "inset(10px round 10px 20px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 10px 20px 10px 20px)") is "inset(10px round 10px 20px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px 20px 30px 40px round 5px 6px)") is "inset(10px 20px 30px 40px round 5px 6px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 20px 30px 40px round 5px 6px)") is "inset(10px 20px 30px 40px round 5px 6px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px 20px 30px 40px round 5px 6px 7px 8px / 50px 60px 70px 80px)") is "inset(10px 20px 30px 40px round 5px 6px 7px 8px / 50px 60px 70px 80px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 20px 30px 40px round 5px 6px 7px 8px / 50px 60px 70px 80px)") is "inset(10px 20px 30px 40px round 5px 6px 7px 8px / 50px 60px 70px 80px)"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 0px / 10px)") is "inset(10px round 0px / 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 0px / 10px)") is "inset(10px round 0px / 10px)"
PASS getCSSText("-webkit-shape-outside", "inset(calc(25%*3 - 10in) 0 0 0)") is "inset(calc(75% - 10in) 0px 0px)"
PASS getComputedStyleValue("-webkit-shape-outside", "inset(calc(25%*3 - 10in) 0 0 0)") is "inset(calc(75% - 960px) 0px 0px)"
PASS getCSSText("-webkit-shape-outside", "circle()") is "circle(at 50% 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle()") is "circle(closest-side at 50% 50%)"
PASS getCSSText("-webkit-shape-outside", "circle(farthest-side)") is "circle(farthest-side at 50% 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(farthest-side)") is "circle(farthest-side at 50% 50%)"
PASS getCSSText("-webkit-shape-outside", "circle(closest-side)") is "circle(closest-side at 50% 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(closest-side)") is "circle(closest-side at 50% 50%)"
PASS getCSSText("-webkit-shape-outside", "circle(10px)") is "circle(10px at 50% 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px)") is "circle(10px at 50% 50%)"
PASS getCSSText("-webkit-shape-outside", "circle(10px at 10px)") is "circle(10px at 10px 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at 10px)") is "circle(10px at 10px 50%)"
PASS getCSSText("-webkit-shape-outside", "circle(10px at 10px 10px)") is "circle(10px at 10px 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at 10px 10px)") is "circle(10px at 10px 10px)"
PASS getCSSText("-webkit-shape-outside", "circle(at 10px)") is "circle(at 10px 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at 10px)") is "circle(closest-side at 10px 50%)"
PASS getCSSText("-webkit-shape-outside", "circle(at 10px 10px)") is "circle(at 10px 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at 10px 10px)") is "circle(closest-side at 10px 10px)"
PASS getCSSText("-webkit-shape-outside", "circle(at top left)") is "circle(at 0% 0%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at top left)") is "circle(closest-side at 0% 0%)"
PASS getCSSText("-webkit-shape-outside", "circle(at right bottom)") is "circle(at 100% 100%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at right bottom)") is "circle(closest-side at 100% 100%)"
PASS getCSSText("-webkit-shape-outside", "circle(10px at left top 10px)") is "circle(10px at 0% 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at left top 10px)") is "circle(10px at 0% 10px)"
PASS getCSSText("-webkit-shape-outside", "circle(10px at top 10px left 10px)") is "circle(10px at 10px 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at top 10px left 10px)") is "circle(10px at 10px 10px)"
PASS getCSSText("-webkit-shape-outside", "circle(10px at right 10% bottom 10%)") is "circle(10px at 90% 90%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at right 10% bottom 10%)") is "circle(10px at 90% 90%)"
PASS getCSSText("-webkit-shape-outside", "circle(10px at right 0px bottom 0px)") is "circle(10px at 100% 100%)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at right 0px bottom 0px)") is "circle(10px at 100% 100%)"
PASS getCSSText("-webkit-shape-outside", "circle(10px at right 10px bottom 10px)") is "circle(10px at right 10px bottom 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at right 10px bottom 10px)") is "circle(10px at right 10px bottom 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse()") is "ellipse(at 50% 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse()") is "ellipse(closest-side closest-side at 50% 50%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px)") is "ellipse(10px at 50% 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px)") is "ellipse(10px closest-side at 50% 50%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 20px)") is "ellipse(10px 20px at 50% 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 20px)") is "ellipse(10px 20px at 50% 50%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px at 10px)") is "ellipse(10px at 10px 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px at 10px)") is "ellipse(10px closest-side at 10px 50%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 20px at 10px)") is "ellipse(10px 20px at 10px 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 20px at 10px)") is "ellipse(10px 20px at 10px 50%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px at 10px 10px)") is "ellipse(10px at 10px 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px at 10px 10px)") is "ellipse(10px closest-side at 10px 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse(at 10px)") is "ellipse(at 10px 50%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(at 10px)") is "ellipse(closest-side closest-side at 10px 50%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(at 10px 10px)") is "ellipse(at 10px 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(at 10px 10px)") is "ellipse(closest-side closest-side at 10px 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse(at top left)") is "ellipse(at 0% 0%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(at top left)") is "ellipse(closest-side closest-side at 0% 0%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(at right bottom)") is "ellipse(at 100% 100%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(at right bottom)") is "ellipse(closest-side closest-side at 100% 100%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px at left top 10px)") is "ellipse(10px at 0% 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px at left top 10px)") is "ellipse(10px closest-side at 0% 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px at top 10px left 10px)") is "ellipse(10px at 10px 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px at top 10px left 10px)") is "ellipse(10px closest-side at 10px 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px at right 10px bottom 10px)") is "ellipse(10px at right 10px bottom 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px at right 10px bottom 10px)") is "ellipse(10px closest-side at right 10px bottom 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 20px at left top 10px)") is "ellipse(10px 20px at 0% 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 20px at left top 10px)") is "ellipse(10px 20px at 0% 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 20px at top 10px left 10px)") is "ellipse(10px 20px at 10px 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 20px at top 10px left 10px)") is "ellipse(10px 20px at 10px 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 20px at right 10px bottom 10px)") is "ellipse(10px 20px at right 10px bottom 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 20px at right 10px bottom 10px)") is "ellipse(10px 20px at right 10px bottom 10px)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 10px at right 10% bottom 10%)") is "ellipse(10px 10px at 90% 90%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 10px at right 10% bottom 10%)") is "ellipse(10px 10px at 90% 90%)"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 10px at right 0px bottom 0px)") is "ellipse(10px 10px at 100% 100%)"
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 10px at right 0px bottom 0px)") is "ellipse(10px 10px at 100% 100%)"
PASS getCSSText("-webkit-shape-outside", "polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("-webkit-shape-outside", "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("-webkit-shape-outside", "polygon(10px calc(10in))") is "polygon(10px calc(10in))"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(10px calc(10in))") is "polygon(10px 960px)"
PASS getCSSText("-webkit-shape-outside", "polygon(calc(10in) 10px)") is "polygon(calc(10in) 10px)"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(calc(10in) 10px)") is "polygon(960px 10px)"
PASS getCSSText("-webkit-shape-outside", "content-box") is "content-box"
PASS getComputedStyleValue("-webkit-shape-outside", "content-box") is "content-box"
PASS getCSSText("-webkit-shape-outside", "padding-box") is "padding-box"
PASS getComputedStyleValue("-webkit-shape-outside", "padding-box") is "padding-box"
PASS getCSSText("-webkit-shape-outside", "border-box") is "border-box"
PASS getComputedStyleValue("-webkit-shape-outside", "border-box") is "border-box"
PASS getCSSText("-webkit-shape-outside", "margin-box") is "margin-box"
PASS getComputedStyleValue("-webkit-shape-outside", "margin-box") is "margin-box"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) content-box") is "polygon(0px 0px, 10px 10px, 10px 0px) content-box"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) content-box") is "polygon(0px 0px, 10px 10px, 10px 0px) content-box"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) padding-box") is "polygon(0px 0px, 10px 10px, 10px 0px) padding-box"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) padding-box") is "polygon(0px 0px, 10px 10px, 10px 0px) padding-box"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) border-box") is "polygon(0px 0px, 10px 10px, 10px 0px) border-box"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) border-box") is "polygon(0px 0px, 10px 10px, 10px 0px) border-box"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) margin-box") is "polygon(0px 0px, 10px 10px, 10px 0px) margin-box"
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) margin-box") is "polygon(0px 0px, 10px 10px, 10px 0px) margin-box"
PASS getCSSText("-webkit-shape-outside", "content-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "content-box polygon(0px 0px, 10px 10px, 10px 0px)"
PASS getComputedStyleValue("-webkit-shape-outside", "content-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(0px 0px, 10px 10px, 10px 0px) content-box"
PASS getCSSText("-webkit-shape-outside", "padding-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "padding-box polygon(0px 0px, 10px 10px, 10px 0px)"
PASS getComputedStyleValue("-webkit-shape-outside", "padding-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(0px 0px, 10px 10px, 10px 0px) padding-box"
PASS getCSSText("-webkit-shape-outside", "border-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "border-box polygon(0px 0px, 10px 10px, 10px 0px)"
PASS getComputedStyleValue("-webkit-shape-outside", "border-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(0px 0px, 10px 10px, 10px 0px) border-box"
PASS getCSSText("-webkit-shape-outside", "margin-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "margin-box polygon(0px 0px, 10px 10px, 10px 0px)"
PASS getComputedStyleValue("-webkit-shape-outside", "margin-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(0px 0px, 10px 10px, 10px 0px) margin-box"
PASS getCSSText("-webkit-shape-outside", "linear-gradient(to right, red, blue)") is "linear-gradient(to right, red, blue)"
PASS getComputedStyleValue("-webkit-shape-outside", "linear-gradient(to right, red, blue)") is "linear-gradient(to right, red, blue)"
PASS getCSSText("-webkit-shape-outside", "radial-gradient(at 10px 20px, red 0%, blue 50%, green 95%)") is "radial-gradient(at 10px 20px, red 0%, blue 50%, green 95%)"
PASS getComputedStyleValue("-webkit-shape-outside", "radial-gradient(at 10px 20px, red 0%, blue 50%, green 95%)") is "radial-gradient(at 10px 20px, red 0%, blue 50%, green 95%)"
PASS removeBaseURL(getCSSText("-webkit-shape-outside", "url('image')")) is "url(image)"
PASS removeBaseURL(getComputedStyleValue("-webkit-shape-outside", "url('image')")) is "url(image)"
PASS getCSSText("-webkit-shape-outside", "calc()") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "calc()") is "none"
PASS getCSSText("-webkit-shape-outside", "auto") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "auto") is "none"
PASS getCSSText("-webkit-shape-outside", "inset()") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset()") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px, 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px, 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px 20px, 30px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 20px, 30px)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px, 20px 30px 40px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px, 20px 30px 40px)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px 20px 30px 40px 50px 60px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px 20px 30px 40px 50px 60px)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(round)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(round)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(round 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(round 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px round)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 10px /)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 10px /)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px round 20px 30px 40px 50px 60px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round 20px 30px 40px 50px 60px)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px round /)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round /)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(10px round / 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(10px round / 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(/)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(/)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(/ 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(/ 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "inset(round /)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "inset(round /)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(10px 20px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px 20px)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(10px at 10px 10px 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at 10px 10px 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(10px at 10px 10px at center)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at 10px 10px at center)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(10px at center center 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at center center 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(10px at 10% left)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at 10% left)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(at 10px 10px 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at 10px 10px 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(at 10px 10px at center)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at 10px 10px at center)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(at center center 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at center center 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(at 10px 10px closest-side)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at 10px 10px closest-side)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(10px at)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(10px at)") is "none"
PASS getCSSText("-webkit-shape-outside", "circle(at)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "circle(at)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 20px 30px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 20px 30px)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px at 10px 10px 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px at 10px 10px 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px at 10px 10px at center)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px at 10px 10px at center)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px at center center 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px at center center 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 10px at 10% left)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 10px at 10% left)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 20px 30px at center center 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 20px 30px at center center 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(at 10px 10px 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(at 10px 10px 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(at 10px 10px at center)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(at 10px 10px at center)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(at center center 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(at center center 10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(10px 20px at)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(10px 20px at)") is "none"
PASS getCSSText("-webkit-shape-outside", "ellipse(at)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "ellipse(at)") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon()") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon()") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero)") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(evenodd)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(evenodd)") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero,10px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero,10px)") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(evenodd,12px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(evenodd,12px)") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(10px, 20px, 30px, 40px, 40px, 50px)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(10px, 20px, 30px, 40px, 40px, 50px)") is "none"
PASS getCSSText("-webkit-shape-outside", "content-box content-box") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "content-box content-box") is "none"
PASS getCSSText("-webkit-shape-outside", "polygon(nonzero, 0 0, 10px 10px, 10px 0) polygon(nonzero, 0 0, 10px 10px, 10px 0)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "polygon(nonzero, 0 0, 10px 10px, 10px 0) polygon(nonzero, 0 0, 10px 10px, 10px 0)") is "none"
PASS getCSSText("-webkit-shape-outside", "content-box polygon(nonzero, 0 0, 10px 10px, 10px 0) content-box") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "content-box polygon(nonzero, 0 0, 10px 10px, 10px 0) content-box") is "none"
PASS getCSSText("-webkit-shape-outside", "none content-box") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "none content-box") is "none"
PASS getCSSText("-webkit-shape-outside", "url('shape.svg') content-box") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "url('shape.svg') content-box") is "none"
PASS getCSSText("-webkit-shape-outside", "url('shape.svg') polygon(nonzero, 0 0, 10px 10px, 10px 0)") is ""
PASS getComputedStyleValue("-webkit-shape-outside", "url('shape.svg') polygon(nonzero, 0 0, 10px 10px, 10px 0)") is "none"
PASS getParentAndChildComputedStylesString("-webkit-shape-outside", "none", "circle(30px at 10px 20px)") is "parent: none, child: circle(30px at 10px 20px)"
PASS getParentAndChildComputedStylesString("-webkit-shape-outside", "circle(30px at 10px 20px)", "initial") is "parent: circle(30px at 10px 20px), child: none"
PASS getParentAndChildComputedStylesString("-webkit-shape-outside", "circle(30px at 10px 20px)", "") is "parent: circle(30px at 10px 20px), child: none"
PASS getParentAndChildComputedStylesString("-webkit-shape-outside", "circle(30px at 10px 20px)", "inherit") is "parent: circle(30px at 10px 20px), child: circle(30px at 10px 20px)"
PASS getParentAndChildComputedStylesString("-webkit-shape-outside", "", "inherit") is "parent: none, child: none"
PASS getParentAndChildComputedStylesString("-webkit-shape-outside", "none", "inherit") is "parent: none, child: none"
PASS successfullyParsed is true

TEST COMPLETE
